Create ArmyGold.cees
