--- Army/Various/ArmySevenCircleWar.cs
@@ -1,67 +1,0 @@
-//cs_include Scripts/CoreBots.cs
-//cs_include Scripts/CoreFarms.cs
-//cs_include Scripts/CoreAdvanced.cs
-//cs_include Scripts/CoreStory.cs
-//cs_include Scripts/Army/CoreArmyLite.cs
-//cs_include Scripts/Story/Legion/SevenCircles(War).cs
-using Skua.Core.Interfaces;
-using Skua.Core.Models.Items;
-using Skua.Core.Models.Quests;
-
-public class ArmySevenCircle
-{
-    private IScriptInterface Bot => IScriptInterface.Instance;
-    private CoreBots Core => CoreBots.Instance;
-    private CoreFarms Farm = new();
-    private CoreAdvanced Adv => new();
-    private CoreArmyLite Army = new();
-    public SevenCircles SC = new();
-
-    private static CoreBots sCore = new();
-    private static CoreArmyLite sArmy = new();
-
-    public string OptionsStorage = "ArmySevenCircle";
-    public bool DontPreconfigure = true;
-    public List<IOption> Options = new List<IOption>()
-    {
-        sArmy.player1,
-        sArmy.player2,
-        sArmy.player3,
-        sArmy.player4,
-        sArmy.player5,
-        sArmy.player6,
-        sArmy.player7,
-        sArmy.packetDelay,
-        sCore.SkipOptions
-    };
-
-    public void ScriptMain(IScriptInterface bot)
-    {
-        Core.SetOptions(disableClassSwap: true);
-        bot.Options.RestPackets = false;
-
-        if (!Bot.Quests.IsUnlocked(7979))
-        {
-            Core.Logger("Doing the storyline");
-            SC.Circles();
-        }
-        Setup();
-        Core.SetOptions(false);
-    }
-
-    public void Setup()
-    {
-        Core.PrivateRooms = true;
-        Core.PrivateRoomNumber = Army.getRoomNr();
-        
-        Core.EquipClass(ClassType.Farm);
-        Core.RegisterQuests(7979, 7980, 7981);
-        Farm.ToggleBoost(BoostType.Gold);
-        Army.SmartAggroMonStart("sevencircleswar", "Wrath Guard", "Heresy Guard", "Violence Guard", "Treachery Guard");
-        while (!Bot.ShouldExit && Bot.Player.Gold < 100000000)
-            Bot.Combat.Attack("*");
-        Army.AggroMonStop(true);
-        Farm.ToggleBoost(BoostType.Gold, false);
-        Core.CancelRegisteredQuests();
-    }
-}
